[PM-9401] Move FlagKey into model folder.
