+++ app/src/main/java/com/x8bit/bitwarden/data/platform/manager/FeatureFlagManagerImpl.kt
@@ -1,6 +1,7 @@
 package com.x8bit.bitwarden.data.platform.manager
 
 import com.x8bit.bitwarden.data.platform.datasource.disk.model.ServerConfig
+import com.x8bit.bitwarden.data.platform.manager.model.FlagKey
 import com.x8bit.bitwarden.data.platform.repository.ServerConfigRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
@@ -55,38 +56,4 @@ private fun <T : Any> ServerConfig?.getFlagValueOrDefault(key: FlagKey<T>): T {
             }
         }
         ?: defaultValue
-}
-
-/**
- * Class to hold feature flag keys.
- * @property [stringValue] corresponds to the string value of a give key
- * @property [defaultValue] corresponds to default value of the flag of type [T]
- */
-sealed class FlagKey<out T : Any> {
-    abstract val stringValue: String
-    abstract val defaultValue: T
-
-    /**
-     * Data object holding the key for Email Verification feature
-     */
-    data object EmailVerification : FlagKey<Boolean>() {
-        override val stringValue: String = "email-verification"
-        override val defaultValue: Boolean = false
-    }
-
-    /**
-     * Data object holding the key for an Int flag to be used in tests
-     */
-    data object DummyInt : FlagKey<Int>() {
-        override val stringValue: String = "email-verification"
-        override val defaultValue: Int = Int.MIN_VALUE
-    }
-
-    /**
-     * Data object holding the key for an String flag to be used in tests
-     */
-    data object DummyString : FlagKey<String>() {
-        override val stringValue: String = "email-verification"
-        override val defaultValue: String = "defaultValue"
-    }
 }
