+++ app/src/test/java/com/x8bit/bitwarden/data/platform/manager/FeatureFlagManagerTest.kt
@@ -5,6 +5,7 @@ import com.x8bit.bitwarden.data.platform.datasource.disk.model.ServerConfig
 import com.x8bit.bitwarden.data.platform.datasource.network.model.ConfigResponseJson
 import com.x8bit.bitwarden.data.platform.datasource.network.model.ConfigResponseJson.EnvironmentJson
 import com.x8bit.bitwarden.data.platform.datasource.network.model.ConfigResponseJson.ServerJson
+import com.x8bit.bitwarden.data.platform.manager.model.FlagKey
 import com.x8bit.bitwarden.data.platform.repository.util.FakeServerConfigRepository
 import kotlinx.coroutines.test.runTest
 import kotlinx.serialization.json.JsonPrimitive
